test(e2e): split points and mistakes tests into separate file
